refactor(debugger): extract object/collection layout into a shared header
Reading an object means knowing where a class keeps its fields and where a
collection keeps its contents. Both debugger front ends need that, and each
carried its own copy, which is how they drifted: the CLI reported a Vector's
backing array instead of its size (#594) and the DAP formatter read a String's
Char[] capacity instead of its length (#598). Each was fixed where it bit, but
the duplication that produced them stayed.

The layout knowledge moves out of DapAdapter into obj_layout.h: FieldIndex,
FindInstanceField, ClassLeafName, CollectionKind, CollectionSize,
CollectionSummary, IsLeafObject, DescribeObject, FormatSlot, ArrayBody,
ElementName, and both FormatVariableValue overloads -- the StackFrame* form
calls the size_t* form, so they travel together. The VarHandleKind enum they
depend on moves with them; the JSON-emitting Expand* functions stay in dap.cpp
and now call the shared layer.

All thirteen are pure functions of the object graph -- no adapter state -- so
the unit is header-only. That is deliberate: dap.cpp is listed in four
Makefiles, the VS project and the Xcode project, and a new .cpp would have to
be added to all six, one of which (the Xcode pbxproj) cannot be verified on
this machine. inline functions in the same Runtime namespace mean no build
file changes and no call-site changes.

Behavior is intentionally unchanged; this commit only moves code. Verified:
debugger CLI 25/25, run_dap_tests.sh 13/13, run_dap_tests.py 20/20.
